+++ ios/ReactNativeBlobUtilFS.mm
@@ -149,8 +149,8 @@ NSMutableDictionary *fileStreams = nil;
 
 + (void) readStream:(NSString *)uri
            encoding:(NSString * )encoding
-         bufferSize:(int)bufferSize
-               tick:(int)tick
+         bufferSize:(double)bufferSize
+               tick:(double)tick
            streamId:(NSString *)streamId
           baseModule:(ReactNativeBlobUtil*)baseModule
 {
